Change expected test result
The singular_route_key() method is used to set the thread ID in emails,
and for some reason this method returns 'commits' for the non-namespaced
Commit class.  But it returns 'gitlab_commit' when the model class is
namespaced, so I'm changing the expected result of an rspec test to
reflect this difference.